Improvments

+++ DownloadUtils.cs
@@ -4,18 +4,24 @@
             if (downloadPreferences == null) { return; }
 
 
-            Utils.RunCommand($"yt-dlp.exe -f {downloadPreferences.FormatString} -o {downloadPreferences.media.ID}.%(ext)s {downloadPreferences.media.URL}", eventOutput);
+            Utils.RunCommand($"yt-dlp.exe -f {downloadPreferences.FormatString} -o {downloadPreferences.media.Guid}.%(ext)s {downloadPreferences.media.URL}", eventOutput);
 
-            if (downloadPreferences.path != null) {
-                if (File.Exists(downloadPreferences.path)) {
-                    File.Delete(downloadPreferences.path);
-                }
+            FileInfo[] filesInfoStartWithGuid = new DirectoryInfo(Directory.GetCurrentDirectory()).GetFiles($"{downloadPreferences.media.Guid}*");
 
-                if (File.Exists($"{downloadPreferences.media.ID}.{downloadPreferences.FileExt}")) {
-                    File.Move($"{downloadPreferences.media.ID}.{downloadPreferences.FileExt}", downloadPreferences.path);
+            if (filesInfoStartWithGuid.Count() != 0)
+            {
+                if (downloadPreferences.path != null) {
+                    if (File.Exists(downloadPreferences.path)) {
+                        File.Delete(downloadPreferences.path);
+                    }
+
+                    string tempFilePath = filesInfoStartWithGuid[0].FullName;
+
+                    if (File.Exists(tempFilePath)) {
+                        File.Move(tempFilePath, downloadPreferences.path);
+                    }
                 }
             }
-
         }
 
     }
